Speaker_Diarization/requirements.txt
whisper
torch
numpy
pydub
scikit-learn
pyannote.audio
pathlib